+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 Unreleased changes.
 
+- [#1174](https://github.com/Zilliqa/zq2/pull/1174): Limit the returned size of `GetSmartContractState` when the `state_rpc_limit` configuration is set.
+- [#1281](https://github.com/Zilliqa/zq2/pull/1281): Emit an `ERROR` level log when a node panics.
+
 ## [0.1.0] - 2024-08-01
 
 Initial release of Zilliqa 2.